+++ packages/swirl-components/src/utils.ts
@@ -176,6 +176,26 @@ export function isMobileViewport() {
   return !window.matchMedia("(min-width: 768px)").matches;
 }
 
+export function parentsPassShadow(node, selector, matches = []) {
+  if (!node) {
+    return matches;
+  }
+
+  if (node instanceof ShadowRoot) {
+    return parentsPassShadow(node.host, selector, matches);
+  }
+
+  if (node instanceof HTMLElement) {
+    if (node.matches(selector)) {
+      return parentsPassShadow(node.parentNode, selector, [...matches, node]);
+    } else {
+      return parentsPassShadow(node.parentNode, selector, matches);
+    }
+  }
+
+  return parentsPassShadow(node.parentNode, selector, matches);
+}
+
 export function querySelectorAllDeep<TargetType extends Element = HTMLElement>(
   root: HTMLElement,
   selector: string